Refactored arg passing tests via TestWiresArgPassingMixin.
